fix: refactor contactFields to staticData and update imports in Keap components
